Add G-Code header comment for center of the board.
This will allow people to pick and choose which programs they want to use instead of having to use them all. Anyone generating, say, their own shape programs will know where the center of the board is, and can adjust G54 accordingly to use the insert or top profile programs.