Overview page generation
Generate a HTML overview page, completely stand-alone, which shows details of the shapes which MonkeyCAM is generating. The basics are there, and more will come soon. Includes a pretty general framework for building these, so adding more should be fast.

Also fixes a bug with insert placement. They were incorrectly placed over the board center, not the effective edge center.