Splitboard support
Added support for splitboards. You can specify a center gap by which the board will be widened down the center so you can cut it in half after layup. A center trench program is added to allow you to inlay sidewall material, etc., down the center of the board. Room is automatically added for edges down the center, and the base cutout program gives half the base (flip the material and run again to get the other half.) Everything else is adjusted as appropriate for the added center gap.

Example output: https://www.monkeycam.org/results/5637369860456448

New board parameter:

* "splitboard": true

New machine parameters:

* "splitboard center trench width": 0.5
* "splitboard center gap": 0.100

This is for issue #11.