Update PubSub error handling
The new worker goes unresponsive now and then. Added more logging, removed old hack for this since it obviously wasn't helping, err on the side of exiting when there is a pubsub error or close.

Also removed the stale http server.

Finally, added a backstop with some simple heartbeat messages.